pyproject.toml
[project]
name = "yolo-vit"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torchvision",
    "timm",
    "numpy",
    "matplotlib",
    "tqdm",
    "pillow",
]
